feat(runtime): extract @excitedjs/agent-runtime-codex (#209)
Issue #209 slice 3: split the built-in Codex runtime out of core into the
publishable package @excitedjs/agent-runtime-codex (alias builtin:codex),
preserving runtime semantics, config, paths, state, tests, and the
server/test factory seams.

The package implements the neutral @excitedjs/dreamux-types
AgentRuntimeProvider and depends on @excitedjs/dreamux-types ONLY — an
import-boundary test rejects any @excitedjs/dreamux import or relative
escape. It owns the Codex engine (supervisor, WS RPC, initialize handshake,
thread start/resume, turn manager, events, completion delivery, config,
args, version gate) and vendors only generic OS/validation/turn helpers,
never host layout/path/socket/log contracts.

builtin/codex/ becomes a thin core-owned adapter: it maps the host create
context (dispatcher row/store/host logger) onto the neutral one and supplies
the host contracts the package must not reconstruct — per-dispatcher paths,
the volatile rendezvous-socket allocator, the package-bin PATH seed, the
dispatcher-store state sink, the Codex home/auth doctor, and the
bundled-skill install. The codex-home doctor, diagnostic, host codex paths,
and re-export shims stay in core so existing import paths are stable.
@excitedjs/dreamux depends on the package by default and resolves
builtin:codex to it.

Deferred to later slices: role-gated skillSources via codex
skills/extraRoots/set and deleting the symlink install (slice 4); the
@excitedjs/agent-runtime-claude-code extraction.
