+++ common/changes/@excitedjs/agent-runtime-codex/i209-extract-codex-runtime_2026-06-13-00-00.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "comment": "Introduce the built-in Codex Agent Runtime package @excitedjs/agent-runtime-codex (alias builtin:codex, issue #209 slice 3). Implements the neutral @excitedjs/dreamux-types AgentRuntimeProvider (Codex app-server supervisor, WS RPC, initialize handshake, thread start/resume, turn manager, teammate completion delivery, config/args/version gate) and depends on @excitedjs/dreamux-types only — never on @excitedjs/dreamux core. Everything host-specific (per-dispatcher paths, the volatile rendezvous-socket root, the durable state sink, the package-bin PATH seed, bundled-skill install, the Codex home/auth doctor) is injected by the host through the neutral create context and provider options.",
+      "type": "minor",
+      "packageName": "@excitedjs/agent-runtime-codex"
+    }
+  ],
+  "packageName": "@excitedjs/agent-runtime-codex",
+  "email": "053700@gmail.com"
+}
