+++ common/changes/@excitedjs/dreamux/i209-extract-codex-runtime_2026-06-13-00-00.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "comment": "Extract the built-in Codex runtime into the publishable @excitedjs/agent-runtime-codex package and depend on it by default (issue #209 slice 3). builtin/codex/ becomes a thin core-owned adapter that maps the host create context onto the neutral @excitedjs/dreamux-types contract and supplies the host path/socket/log/state/skill contracts the package must not reconstruct; the codex-home doctor, diagnostic, host paths, and re-export shims stay in core so existing import paths are stable. The builtin:codex alias, runtime semantics, config, paths, and state are unchanged; no operator action required on upgrade.",
+      "type": "patch",
+      "packageName": "@excitedjs/dreamux"
+    }
+  ],
+  "packageName": "@excitedjs/dreamux",
+  "email": "053700@gmail.com"
+}
