+++ packages/agent-runtime/codex/src/index.ts
@@ -1,0 +1,72 @@
+/**
+ * `@excitedjs/agent-runtime-codex` — the built-in Codex Agent Runtime provider
+ * for Dreamux (alias `builtin:codex`). Implements the
+ * `@excitedjs/dreamux-types` `AgentRuntimeProvider` contract; never imports
+ * `@excitedjs/dreamux` core.
+ */
+
+export {
+  createCodexAgentRuntimeProvider,
+  resolveCodexBinPath,
+  codexRuntimeArgsForMcpServers,
+  dispatcherCodexConfig,
+  CODEX_AGENT_RUNTIME_CAPABILITIES,
+  type CodexAgentRuntimeProviderOptions,
+} from './provider.js';
+
+export {
+  CodexRuntime,
+  type CodexRuntimeDeps,
+  type CodexWorkspaceSkillPrepResult,
+} from './runtime.js';
+
+export {
+  CodexProcess,
+  type CodexProcessOptions,
+  type CodexProcessExit,
+} from './supervisor.js';
+
+export {
+  CodexWsClient,
+  type CodexWsClientOptions,
+  type NotificationHandler,
+} from './rpc.js';
+
+export type {
+  ServerNotification,
+  ServerRequest,
+  ThreadStartResponse,
+  TurnStartResponse,
+} from './types.js';
+
+export { performInitializeHandshake } from './handshake.js';
+
+export { codexMcpServerArgs } from './mcp-config.js';
+
+export {
+  parseCodexArgs,
+  codexArgsFromConfig,
+  codexArgsToCli,
+  type ParsedCodexArgs,
+} from './args.js';
+
+export {
+  type DispatcherCodexConfig,
+  readDispatcherCodexConfig,
+  defaultDispatcherCodexConfig,
+  DEFAULT_CODEX_BIN,
+  DEFAULT_INITIALIZE_TIMEOUT_MS,
+  DEFAULT_CODEX_TURN_TIMEOUT_MS,
+  DEFAULT_APPROVAL_POLICY,
+  DEFAULT_SANDBOX_MODE,
+  ALLOWED_APPROVAL_POLICIES,
+  ALLOWED_SANDBOX_MODES,
+} from './config.js';
+
+export {
+  MIN_CODEX_VERSION,
+  parseCodexVersion,
+  codexVersionSatisfies,
+} from './version.js';
+
+export { BUILTIN_CODEX_PROVIDER_REF } from './provider-ref.js';
