+++ packages/dreamux/CLAUDE.md
@@ -21,7 +21,7 @@ Two settled shape rules govern where code lives:
 |---|---|---|
 | `server.ts` | process entry + wiring only | builds registry/catalog/store/services, opens the admin socket, starts dispatchers; owns no teammate/channel/runtime orchestration |
 | `agent-runtime/` | the AgentRuntime seam: contract (`types.ts`, `turn.ts`), `catalog.ts`, `external-provider.ts` loader, `load-config.ts` (caller-composed builtin registration), `index.ts` barrel | one runtime abstraction for every agent role — see [`agent-runtime/CLAUDE.md`](src/agent-runtime/CLAUDE.md) |
-| `agent-runtime/builtin/codex/` | the whole `builtin:codex` stack: supervisor/rpc/events/handshake/types (transport), runtime, provider, args, codex-home, paths, turn-manager, approval | codex specifics close over here; nothing codex-named leaks out |
+| `agent-runtime/builtin/codex/` | the core-owned `builtin:codex` **adapter** over the published `@excitedjs/agent-runtime-codex` package (#209 slice 3): `provider.ts` (host→neutral adapter), `paths.ts`/`codex-home.ts`/`diagnostic.ts`/`runtime-support.ts` (host contracts), and re-export shims; the engine itself (supervisor/rpc/handshake/turn-manager/events/runtime/args/config) lives in the package | codex specifics close over here or in the package; nothing codex-named leaks out, and the package never imports core |
 | `agent-runtime/builtin/claude-code/` | the whole `builtin:claude-code` stack: supervisor/rpc/stream/types, runtime, args, paths, mcp-config | claude specifics close over here |
 | `dispatcher-service/` | the Dispatcher Service entity — see [`dispatcher-service/CLAUDE.md`](src/dispatcher-service/CLAUDE.md) | holds the dispatcher agent + orchestrates teammates |
 | `dispatcher-service/dispatcher/` | `DispatcherAgentService` (slots / start / resume / stop / restart-notice / channel session / role MCP injection) + dispatcher base prompt | dispatcher agent lifecycle is tied to the server |
