+++ packages/dreamux/src/agent-runtime/builtin/codex/mcp-config.ts
@@ -1,9 +1,20 @@
-import type { AgentRuntimeMcpServer } from '../../types.js';
+/**
+ * Core-side Codex MCP bridge.
+ *
+ * The generic `codexMcpServerArgs` renderer now lives in
+ * `@excitedjs/agent-runtime-codex`; it is re-exported here so existing core/test
+ * import paths stay stable (issue #209 slice 3). The Feishu-specific helper
+ * stays in core: it crosses the channel boundary (the Codex runtime package must
+ * not know about Feishu), so core wires the Feishu MCP descriptor into the
+ * generic renderer.
+ */
+import { codexMcpServerArgs } from '@excitedjs/agent-runtime-codex';
 import {
   feishuMcpServerDescriptor,
   type FeishuMcpServerDescriptorOptions,
 } from '../../../channel/feishu/feishu-mcp-surface.js';
 
+export { codexMcpServerArgs };
 export {
   FEISHU_MCP_SERVER_NAME,
   feishuMcpServerDescriptor,
@@ -14,23 +25,4 @@ export function feishuMcpCodexArgs(
   opts: FeishuMcpServerDescriptorOptions,
 ): string[] {
   return codexMcpServerArgs([feishuMcpServerDescriptor(opts)]);
-}
-
-export function codexMcpServerArgs(
-  servers: readonly AgentRuntimeMcpServer[],
-): string[] {
-  return servers.flatMap((server) => [
-    '-c',
-    `mcp_servers.${server.name}.command=${tomlString(server.command)}`,
-    '-c',
-    `mcp_servers.${server.name}.args=${tomlStringArray(server.args)}`,
-  ]);
-}
-
-function tomlStringArray(values: string[]): string {
-  return `[${values.map(tomlString).join(', ')}]`;
-}
-
-function tomlString(value: string): string {
-  return JSON.stringify(value);
 }
